Implementation of repo_upload_packfile_base_item endpoint
Summary: The previous diff in the stack introduced the `repo_upload_packfile_base_item` endpoint and this diff adds the relevant logic in that method. It also updates `remote-gitimport` to make use of this endpoint replacing the stub that existed in its place. This will be E2E tested once I use this new uploader method in gitimport and remote-gitimport.

Reviewed By: markbt

Differential Revision: D52164963

fbshipit-source-id: a73ae8a3297b0df34f44702db1a750cb04ed29ab