commandstate: don't require the wlock
Summary:
We've had a lot of issues related to the "undo" extension now suddenly taking the wlock via <pyworkingcopy>.commandstate().

In particular:
- Commands that take store lock but not wlock can get "out of order" lock errors if the "undo" extension ends up taking the wlock "later".
- Commands such as "continue" that re-enter command dispatch can deadlock if both the outer and inner command cause "undo" to take the wlock (because the "repo" object is not shared between the commands).

Fix for now by no longer requiring the wlock for sniffing the repo's command state. The mean reason I wanted to take the wlock was because the repo command state can be spread across multiple files. Probably the long term solution is to conoslidate the state into a single file that is updated atomically, so reading is always consistent.

Reviewed By: zzl0

Differential Revision: D52267123

fbshipit-source-id: 5deab4c30628032c3591ea889262e6d2a4778a9a