Add "highlighted_rev" info to templater
Summary:
Pass the hex value of the "current selected" rev in `hg sl -i` to the templater so that we can display the rev differently using the templating engine.
This will get passed as part of `mapping` in the template engine.

Also adds a templatealias and color configs to show the selected commit as cyan

Reviewed By: quark-zju

Differential Revision: D52178017

fbshipit-source-id: eb691d450ff5b5da60b7970360c491df5f7a0111