Allow walker to bypass incorrect hg changeset blobs
Summary: Related to the S382653, where a commit with author name including the newline character became a public commit. Although the team rewinded the bookmark and then included the amended commit with the fixed name, the older commit with the incorrect author name still exists as a draft commit. When walker comes across this commit, it tries to parse its hgchangeset and invariably fails. This diff tries to add a bypass in case of such commits.

Reviewed By: mitrandir77

Differential Revision: D52256844

fbshipit-source-id: cfbaa0bc4f47a9f1fba01ecbf47367fae9480297