disallow sparse commands in edensparse repos
Summary:
While the `sparse` extension technically shouldn't be enabled while FilteredFS is in use, it is possible a user can invoke a command with `--config extensions.sparse=` to manually enable it (or they could even enable the extension in their .hgrc).

If someone manages to do this, we should warn them that they are using edensparse and don't need to run sparse commands.

Reviewed By: muirdm

Differential Revision: D51970398

fbshipit-source-id: ea941f9563f6e87d02728c1425a96d73394ec0dc