add more varint utils
Summary:
When using FilteredBackingStores, Eden now stores FilteredRootIds in the Snapshot file. FilteredRootIds consist of the original underlying BackingStore's RootId, the length of the FilteredRootId (in varint format), and the filter that was applied to the repo during the last checkout.

To properly parse out the original underlying RootId (in most cases a Mercurial commit hash), we need to parse the varint that describes the length of the underlying RootId and parse the Snapshot file accordingly.

This diff introduces the utils that will be used to do that varint parsing logic.

Reviewed By: kmancini

Differential Revision: D51926558

fbshipit-source-id: f35ba4af4cf59e593fe86b4411d7f94cd722946d