fix renderRootId behavior for FilteredBackingStores
Summary: The intent of RenderObjectId and RenderRootId is that it returns a string that's "human readable" or sane for a caller to examine. Returning a FilteredObjectId or FilteredRootId isn't really the correct thing to do here (callers don't care about the internal details of the FilteredBackingStore). Instead, we should render the Object/Root Id from the underlying BackingStore since that's what most callers would expect.

Reviewed By: kmancini

Differential Revision: D51926553

fbshipit-source-id: 1e23c2b9501a75d0b817e60b6f698e3b821f2551