add `bot()` revset as alias for `bottom()`
Summary:
I tend to use `bot` as the antonym for `top` while moving through commits with sapling, so I made this to add `bot` as essentially an alias for the `bottom` revset.

If this is too much to maintain or there's a better way to alias a revset without defining a new function, let me know!

Reviewed By: quark-zju

Differential Revision: D51208815

fbshipit-source-id: 44c25c943c038e36340543219e4b11c244effddb