pass CREATE_NO_WINDOW and STARTF_USESHOWWINDOW to ExternalTelemetryLogger log_sample
Summary:
`edenfsctl_config_manager` was previously causing cmd window to flash on Windows: https://fb.workplace.com/groups/edenfswindows/posts/1414741589441715

I attempted to use a subprocess like how we do for `edenfsctl start --if-necessary` and `CoCo`, but it was still popping up a window. I missed this instance of `subprocess`, which looks like it still causes the command window to flash.

Testing the config manager with this change gets rid of the command window flashing.

Reviewed By: MichaelCuevas

Differential Revision: D52371460

fbshipit-source-id: a1e791f122d44a7a00063b5ab33949d25fb7055a