Ignore invalid bonsai changesets encountered during walk
Summary:
This is related to the same SEV where a commit with a newline in the author name somehow slipped in. Although we rewound the bookmark to include the amended and fixedx version of the commit in the main branch, there are still draft commits that exist with the incorrect author name. Most production services do not care about the draft commit, except walker which walks even draft commits now.

I already landed a diff which allowed walker to ignore invalid hgchangesets but then it started failing at invalid bonsai changesets. This diff adds a bypass for bonsai changesets so walker can proceed without crashing.

Reviewed By: mitrandir77

Differential Revision: D52388712

fbshipit-source-id: d26933432ef1d0907c31313b734037deaf6becde