Remove dependency on object entry for fetching raw contents of git object
Summary: The `get_object_bytes` function fetches the content of raw git objects from Mononoke data store. Since blobs are not stored directly as raw objects, it creates the git object from the raw content of the blob stored as file content in Mononoke. For other types of objects, it directly fetches the raw git bytes. This function is reusable outside the context of `GitDeltaManifest` and hence removing the dependency on `ObjectEntry` (which is anyway redundant) makes the function more generic and usable in different context (later in the stack)

Reviewed By: markbt

Differential Revision: D52421784

fbshipit-source-id: c7f09126877513e89d50fea8c8e16286ad6e163d