Update gitbundle create admin command to support PackfileItemInclusion enum
Summary: With all the building blocks in place, we can finally add support to the `gitbundle create` newadmin command to either ignore, fetch or fetch and store the base packfile items used while constructing the packfile/bundle for the given repo. If the clone performance using the newadmin command improves with this change, we would default to fetching instead of generating the packfile items in Git protocol as well.

Reviewed By: markbt

Differential Revision: D52443890

fbshipit-source-id: e77558c51c2dcdd1a50e4ed125ff33d631a436d8