Honor PackfileItemInclusion while generating packfile stream
Summary:
The previous diff introduced the `PackfileItemInclusion` enum as a way for the caller to specify whether their packfile items for base git objects be generated or fetched. This diff implements the actual support for this enum in `generator.rs`. Note that for now all callsites choose to remain consistent with the existing setting (i.e. `PackfileItemInclusion::Generate` which represents the state before this enum was introduced). The existing integration tests will show if there has been any regression.

Later in the stack I will add an option in the bundle creator newadmin command to allow the caller to specify the variant they want to use which will be tested by its own integration test.

Reviewed By: markbt

Differential Revision: D52429440

fbshipit-source-id: 41c7e193da2c934edb7fab810efbc67dbecea706