Introduce PackfileItemInclusion enum
Summary:
Currently when generating the `PackfileItem` stream that feeds into the actual packfile/bundle, we fetch either the delta object or the raw git object, then convert them into `PackfileItem` which is basically a no-op for the delta object case but involves encoding and transformation for the raw git object case. This encoding and transformation can quicly become a bottleneck when generating the actual bundle. This diff introduces the `PackfileItemInclusion` enum which will be user exposed to allow the user to specify if they want to:
- Fetch direct packfile items if they exist
- Generate packfile item from raw objects regardless of whether they exist
- Fetch direct packfile items if they exists, otherwise generate and store them

Reviewed By: markbt

Differential Revision: D52422284

fbshipit-source-id: 5da52065673300d172fbf879329a63af40c7d4b8