rendezvous: move delay and threshold into options
Summary:
In D52200140 I removed `TunablesRendezVousController` in favour of always using configured values.  This had the side-effect of switching tests to use the default values for rendez-vous, instead of the implicit zeros that came from unconfigured tunable values.  Since tests also set `free_connections` to 0, this meant every SQL request would wait for 5ms (or 50 concurrent requests, which is unlikely in a test).  This slowed down some of the tests to the point that they timed out.

To fix this, we move the values for `max_delay` and `max_threshold` into `RendezVousOptions`.  This lets us extend `RendezVousOptions::for_test` to use the same values as the tests were using previously.  We keep the default values in prod by defining `Default` for `RendezVousOptions`.

Reviewed By: RajivTS, mzr

Differential Revision: D52430875

fbshipit-source-id: 89f2f75839c598b16645906ef29e7b0bf0347318