dag: port Rust renderdag to TypeScript
Summary:
Currently, ISL converts the DAG into "tree"s and rely on the browser's layout
engine to draw the trees. It has a few issues:
- Cannot draw merges.
- Can be tricky to deal with non-continuous commits (but still doable).

This diff ports the Rust `renderdag` library to TypeScript so we get a proper
engine to figure out the columns, edges given a subset of a graph.

I thought about Rust->WASM. However, it is still annoying with webpack (cannot
use regular `import` at the top of the files, all related logic has to be
`async`). The core logic is about ~800 lines so translation seems feasible.
CodeCompose helped part of the translation when given a temporary comment block
like "translated from Rust: <part of rust code>".

Some of the features like handling long messages are unnecessary in browser,
but might still be useful for tests.

Reviewed By: muirdm

Differential Revision: D52338507

fbshipit-source-id: d61e83f34c37d280f14b63262452e14e25be2669