[reviewstack] Updating textmate build to append relative import .js prefix
With more current versions of node, the `--experimental-specifier-resolution=node'` does not work anymore and running `yarn codegen` from the `reviewstack` folder will fail because of textmate module.

This PR adds a step in the `addons/reviewstack/textmate.js` to fix textmate imports to use .js extension. This will only impact the transpiled .js files and not the source .ts files