Adding concurrency in generating the packfile stream
Summary: Based on Mark's work of trying to add concurrency in the bundle generation logic, this diff sprinkles a bunch of `try_buffered(_)` that reduces the overall time taken for packfile stream generation. Before this change, cloning of aosp/manifest repo for the main branch took ~1min25secs. After this change, it takes 40secs.

Reviewed By: markbt

Differential Revision: D52445789

fbshipit-source-id: 313f77392ba3b26a7a85fca9660dab5b218d6925