split: update split to work better with diff shifting
Summary:
The diff shifting introduces a new case for the split, imagine:

    | | | |function a() {
    | | | |  ...
    | | | |}
    |+| | |
    |+|+|<|function b() {
    |+|+|<|  ...
    |+|+|<|}
    | |+|<|
    | | | |function c() {
    | | | |  ...
    | | | |}
     ^ ^ ^
     | | "move left" button shown
     | displayed in split UI (no shifting)
     internal in linelog (due to shifting, see previous change)

When clicking the "move left" button of the empty line, the line internally in
linelog exists both in aRev and bRev, so the current `add(aRev)` approach does
nothing. To properly "move" the line left, we can insert a line that only exists
in aRev (so it is an insertion compare to aRev-1, and bRev already has that
line so we don't need to duplicate it there).

For "move insertion right", the current logic should work regardless, since
diff shifting only affects `aRev`, logic that only uses `bRev` (like "move
insertion right") should work as-is. In theory the "move deletion left" also
needs change. But we don't do diff shifting for deletion so it'll be okay.

Reviewed By: evangrayk

Differential Revision: D52514622

fbshipit-source-id: d605b242985c35af4896fa96f712decb50cc3df9