linelog: optimize dependency by shifting diff blocks
Summary:
See the added comment for what is going on.

I originally implemented it as splitting out blank lines so

  +(empty line)
  +func {
  +  ...
  +}

becomes 2 insertions:

  +(empty line)

and:

  +func {
  +  ...
  +}

However, I like the current solution more. It does not depend on the line
contents. Cases like `//`, `#` prefix would benefit from this change. It
requires a change for split, which will be fixed in the next diff.

Reviewed By: evangrayk

Differential Revision: D52514618

fbshipit-source-id: 0994742d0a502fa2157848391b6d4f77c818c2eb