rebase: support --rev --dest pairs to specify different destinations
Summary:
For example,

  rebase -r A -d B -r C -d D -r E+F -d G

will rebase A to B; C to D; E and F to G.

This makes it easier to "encode" multiple rebases with different destinations
in one command (which works with --continue and --abort). It is intended to
replace ISL "queued rebases" which can easily be aborted when there is a
conflict.

Note the multi-dest rebase feature already exists and is well tested.
It was just not exposed as an easy-to-use CLI. The test used a customzied
`map` function to use it. But most users won't be able to define such
revset functions.

Reviewed By: zzl0

Differential Revision: D52514616

fbshipit-source-id: 60737d2ec97cb95ff00c8c8804b09aa196225911