Remove disable_commit_scribe_logging_scs tunable
Summary:
The tunable `disable_commit_scribe_logging_scs` was added by stas in D25558248 when we initially started logging commits as part of bookmark movement/creation in SCS. The reason for adding this tunable was to allow an option to switch off this new logging in case it involved a large number of commits which would be expensive to log. Since this logging takes place as part of the request, this could end up increasing the overall latency of the request.

Its been 3-4 years since this tunable was introduced and it looks like logging the commits as part of SCS bookmark movement/creation is not that expensive. Hence, getting rid of this tunable.

Reviewed By: YousefSalama

Differential Revision: D52597618

fbshipit-source-id: 5eadd0f17ded5078d0b3849e81a0b4f717cd1a3d