Replace disable_repo_client_warm_bookmarks_cache tunable with JK
Summary: The tunable `disable_repo_client_warm_bookmarks_cache` was added to act as a killswitch in case we want to disable warm bookmark cache for `RepoClient`. There is an option for this in RepoConfig as well but this tunable acts as an overriding killswitch. We still might want to use this in the future hence moving it to a JK

Reviewed By: YousefSalama

Differential Revision: D52598777

fbshipit-source-id: f5793904c571b43b62e6f22b2fd2b3c742071cf2