tunables: finish deleting 78 tunables
Summary:
As a team, we have been migrating our tunables to jk or deleting them alltogether.

See [this spreadsheet](https://docs.google.com/spreadsheets/d/18DZy1QNBs_raujS-8ibsDjIbB8FlFrA1s8oAoTGRtYA/edit#gid=0) which we use to coordinate.

To avoid conflicts, we haven't actually deleted each tunable from the list in the code.

Now is a good quiet time between hack-a-thons to do a bit of cleanup.

This means that the code is now much closer to a source of truth again, giving better visibility into what actually remains.

This deletes 78 now unused tunables, leaving 26 tunables that are still actually used in the code.

Reviewed By: clara-9

Differential Revision: D52629281

fbshipit-source-id: 484fcd79fb338902d55d8bb4180e2ce69de2fa08