filesystem: kill pendingchanges() in Python
Summary:
pendingchanges() powers "status", but that is handled in Rust now, so we can clean this up.

There a still a few other random things the "fsmonitor" Python extension does, so we can't remove it completely.

I killed the "detectrace" and "walk_on_invalidate" fsmonitor config knobs. They aren't supported in Rust and I don't think we need to add them.

Reviewed By: quark-zju

Differential Revision: D52492697

fbshipit-source-id: 2ebfa907692e56870de5a90208481205d0272764