Make interactiveui.render return a list
Summary: Make the `render` method in the `viewframe` class return a list of strings representing lines to display. This gives us more control over how we display and moves the `splitlines` call to the caller. This will allow us to do things like 'scroll' the display to a certain line later.

Reviewed By: muirdm

Differential Revision: D52496835

fbshipit-source-id: 121be290a91eefbdb3f1f2d8cface5b0e456676d