status: skip nested ".hg" files w/ watchman
Summary:
"status" is supposed to ignore nested repos. There are two cases:

1. A nested working copy file (e.g. "foo/nested_repo/file").
2. A nested repo file (e.g. "foo/nested_repo/.hg/dirstate").

Previously we filtered neither in the watchman case. In this diff I add filtering for 2) since that is relatively easy. Filtering for 1) is less simple since we don't know what the nested repo roots are. We would either have to do a separate watchman query for directories named ".hg" and filter those out in the second query, or post-filter each watchman result by walking up the directory ancestors looking for non-root ".hg" directory presence.

Reviewed By: quark-zju

Differential Revision: D52695803

fbshipit-source-id: 339049373cbb45621144a9706add3a3fa8cda0ae