add optimization for phases
Summary:
fix optimization for phases

this optimisation in the land_stack codepath is currently missing comparing to the unbundle codepath.

the optimisation is coming from:

https://www.internalfb.com/code/fbsource/fbcode/eden/mononoke/repo_client/unbundle/src/processing.rs?lines=348

By moving the phases update into the generic code, we make sure that landing to a public bookmark updates the phases regardless of the codepath used.

In practice, it will now happen on Land Service side that should be beneficial for performance.

Reviewed By: zzl0

Differential Revision: D52729660

fbshipit-source-id: fea7e19872d174818a8450a32eebd4dc5ef7ae2d