status: get rid of "last_write" argument
Summary:
This argument shouldn't be in the application-facing API since the application doesn't know what the "last write" time is.

In general the treestate.invalidate_mtime() mechanism deals with this problem. When writing out the treestate, entries whose mtime is "now" are set as NEED_CHECK so "status" can detect future modifications within the same mtime window.

In this diff I added another call to invalidate_mtime() in the status code path to handle the case where you haven't written out the dirstate yet. This was needed to fix test-import-merge.t.

Reviewed By: quark-zju

Differential Revision: D52424690

fbshipit-source-id: e23c578265e64513fa45fb15ecfda203d665bdad