repo: kill "ensure_locked" API
Summary:
This no-longer-used API encourages adhoc runtime locking which is hard to maintain. Instead, we should follow the Rusty pattern where locking yields a privileged object.

We may still need something similar to this in the Python bindings to allow conversion between a Python-acquired lock and a Rust privileged object.

Reviewed By: quark-zju

Differential Revision: D52424689

fbshipit-source-id: 81edab3e09b860fb6212aca53f35aef2a212dbd9