kill the "testlib" test framework
Summary:
testlib was a new pure-Python integration test framework introduced relatively recently. It had these advantages:
- use "modern" configs by default
- support running against EdenFS and Mononoke (theoretically)
- treat integration tests as code rather than CLI transcript style

However, it had the critical disadvantage of being a second system when the first system (i.e. dot-t tests) was "good enough" and was widely used in a thousand tests. Considering that all of the advantages above can be implemented in dot-t tests, I don't think it is worth it to live in a world with multiple competing test frameworks.

I previously migrated the useful testlib tests back to dot-t tests. In this diff I removed the clone/ tests - they weren't that useful and are already covered by dot-t tests such as test-sparse-clone.t.

Reviewed By: sggutier

Differential Revision: D52669623

fbshipit-source-id: 544a5ef41ce7db22f708033bf3c4b51a6df66125