treestate: track longest prefix when finding filtered keys
Summary:
Make get_filtered_key also return the longest directory prefix encountered. I'm going to use this to avoid dirstate directory divergence on case insensitive filesystems.

For example, if treestate is tracking "foo/A" and we see an untracked file "FOO/B", we want to add it to treestate as "foo/B" to avoid case insensitive "foo"/"FOO" divergence.

Reviewed By: zzl0

Differential Revision: D52669620

fbshipit-source-id: 3dc0a686155b698726747d1279dd58c0e5b5b750