status: fix case insensitive edge case
Summary:
In porting the test status/icasefs.py back to a dot-t test, I found that the Rust status still has some case insensitive behavior mismatches. Taking advantage of improved path normalization of previous diff, I tweaked "status" to normalize untracked files' directories to match the case of directories of existing tracked files in the treestate.

Most of the BUGBUG notes from status/icasefs.py are now resolved. There is one remaining discrepancy, but I don't think it is a big deal (and it was how it worked before).

Reviewed By: liubov-dmitrieva

Differential Revision: D52669622

fbshipit-source-id: cb6e02a0df305f0a9aef66d68d0add5f80d40541