treestate: make normalization work for untracked files
Summary:
On case insensitive filesystems, we want to avoid diverging the treestate with paths such as "foo/a" and "FOO/b".

This can crop up when directory "foo" is renamed on disk to "FOO", and then "b" is created in it. Since "b" wasn't in the treestate, our path normalization would previously not normalize "FOO/b" to "foo/b". In this diff, I make normalization take into account the longest directory prefix so it can normalize "FOO/b" to "foo/b", even though "foo/b" doesn't exist in the treestate yet.

Reviewed By: zzl0

Differential Revision: D52669621

fbshipit-source-id: a3ae2ca16a02f5c5406954085107fa374b7b3b91