dag: warn slow find_free_span
Summary:
It turns out the `find_free_span` is slow. So let's warn about the long loops.

How to find that `find_free_span` is slow? I used [`uftrace`](https://github.com/namhyung/uftrace).
It's a bit hard to configure, so I wrote what I have done:

Switch to the nightly Rust compiler, which supports `-Z` flags:

  % rustup override set nightly
  % make local # regenerate .cargo/config.toml to use nightly Rust

Build the benchmark program with on `-Zinstrument-mcount` and its related flags
(LTO is incompatible with mcount, and we need to set the LLVM pass explicitly
to make it effective):

  % export RUSTFLAGS='-Zinstrument-mcount -Cpasses=ee-instrument<post-inline> -Clinker-plugin-lto=off -Clto=off'

Record a trace (note the program can run 100x slower so I run the benchmark on
the default Mozilla graph, not the production graph):

  % uftrace record -F 'dag*' cargo-target/release/dag-benches import_pull

Analyze the trace, filtering out functions that take longer than 1s:

  % uftrace graph -t 1s
  # Function Call Graph for 'dag-benches' (session: 2c924880fb7d2872)
  ========== FUNCTION CALL GRAPH ==========
  # TOTAL TIME   FUNCTION
      1.011  m : (1) dag-benches
      1.011  m : (1) dag_benches::main
      1.011  m : (1) dag_benches::dag_import::main
      1.010  m : (1) minibench::bench
      1.010  m : (1) dag_benches::dag_import::main::_{{closure}}
      1.010  m : (1) _<minibench..measure..WallClock as minibench..measure..Measure>::measure
      1.010  m : (1) nonblocking::non_blocking_result
      1.010  m : (1) _<core..pin..Pin<P> as core..future..future..Future>::poll
      1.010  m : (1) _<dag..namedag..AbstractNameDag<dag..iddag..IdDag<IS>,M,P,S> as dag..ops..DagImportPullData>::import_pull_data::_{{closure}}
     47.092  s :  +-(40) dag::namedag::find_free_span
               :  |
      3.019  s :  +-(1) dag::iddag::IdDag<Store>::build_segments_from_prepared_flat_segments
               :  |
      1.441  s :  +-(1) dag::namedag::AbstractNameDag<dag..iddag..IdDag<IS>,M,P,S>::persist

Obviously, there are 40 `find_free_span` calls that take 47s in total.
`find_free_span` is terribly slow.

I also tried the `-Zinstrument-xray`. But it seems it does not play well with
`uftrace` - function exits are not detected and the program can crash with
illegal instruction.

Example run with the logging enabled:

  $ LOG=dag=warn cargo run --release -- 'import_pull'
  WARN PERF: find_free_span took too long count=9625 low=2443 reserve_size=1 covered=0..=12066
  WARN PERF: find_free_span took too long count=4435 low=15610 reserve_size=1 covered=0..=20043
  WARN PERF: find_free_span took too long count=8630 low=20131 reserve_size=1 covered=0..=28759
  ...
  WARN PERF: find_free_span took too long count=332440 low=0 reserve_size=1 covered=0..=332438
  ... (~400 lines)
  import_pull_data                                    1.550 s

Obviously, `find_free_span` can be made smarter, which will be fixed by the next change.

Reviewed By: zzl0

Differential Revision: D52758306

fbshipit-source-id: 146b03f1184cbced3bdd5b9c9439120e1b63199e