pycbor: bindings to serialize and deserialize in CBOR format
Summary:
Expose `serde_cbor` features to Python. This can be useful to serialize things
that cannot be represented in JSON, like numbers in map keys or non-utf8 bytes.
With `cpython-ext` this can be implemented easily.

There is a `utils.cborutil`. However it does not handle `str` types.

I plan to use this to dump a production payload of the "clone data" for
investigation.

Reviewed By: markbt

Differential Revision: D52758311

fbshipit-source-id: e0c1e32a8c679239afe7dc370368d2b49aa41b9e