dag: introduce "You are here" virtual commit
Summary:
The "You are here" thing can actually be treated as a commit and get involved
into operations. In revsetlang, "You are here" is "wdir()", different from "."
(working parent, or "HEAD" in Git terminology).

By treating "wdir()" as a separate commit in the dag, certain things might make
sense:
- Dragging "wdir()" means "goto", but can have a UX similar to DnD rebase.
- "Combine" "wdir()" with parent means amend.

Related, one selling point of `jj` is that it always treat `wdir()` as a
separate commit.

Reviewed By: evangrayk

Differential Revision: D52546780

fbshipit-source-id: 666b9dbaa7f443b9cb7d5b72b98e28bbb2aa2bb2