dag/render: provide information to follow lines
Summary:
In ISL, we want to use a different color to draw the "You are here" virtual
commit. The current rendering logic does not yet provide information about
which lines / edges are originated from which commit. For example, in this case:

      │ (You are here)
      ├─╯
      o

We want to use a different color to draw these edges:

        (You are here)
      ╭─╯
      o

Note the "╭" and the "├" belong to a same `LinkLine` unit so we need to specify
colors for subsets of `LinkLine`s.

This diff changes the diff rendering logic to report:

- What "subsets" of `LinkLine`s are from the commit being rendered.
- What column indexes before (nodeColumn) and after (parentColumns) the link
  line are from the commit being rendered.

This will provide enough information for the "You are here" color case, and
potentially other use-cases too. Check the added tests that visualize the
added information.

Reviewed By: muirdm

Differential Revision: D52558936

fbshipit-source-id: 47861e04475288777ace962274a7e19730049a68