dag/render: provide isHead isRoot information
Summary:
In CLI rendering we don't need to distinguish between these two on the "o"
line, since "o" takes the entire height of that row:

                 |
    o            o
    |            |

However, in the browser rendering the "o" does not take all the height so
we need to know whether the vertical line should cover the upper half.
Similarly, we also want to know if the node has parents.

This diff adds `isHead` and `isRoot` fields to answer that.

Reviewed By: evangrayk

Differential Revision: D52546781

fbshipit-source-id: d3a4961a1d2c7deaea757caa672f9e546be6d9af