RenderDag: enable animation
Summary:
Get animation with just 4-line change!

There are feedback about animation (ex. fading in or out the menu) could
hurt productivity. In this case, it's a sliding, not fading animation, making
it a bit easier to follow what are changed / moved / not changed. So it seems
to provide some value.

With this, for example, it seems to make sense to avoid double "You were here",
"You are moving here" labels, since the moving source is animated to destination.

Reviewed By: muirdm, evangrayk

Differential Revision: D52667620

fbshipit-source-id: e241454ca0d04d4d51ec307b293881ac1e97bba4