AnimatedReorderGroup: ignore small distance animation to workaround VSCode button rendering
Summary:
I notice that when committing, while the optimistic state matches the final
state, the commits below moves up and down unnecessarily.

Upon debugging using a breakpoint in AnimatedReorderGroup when it decides to
animate, I found a 4px animation, and the VSCode buttons [Uncommit] [Split]
were in its "primary" style, with a larger height first rendered. Apparently
the VSCode button then change to the specified "icon" style, with a different
height, but it does not do that in the first render. There are enough pitfalls
in VSCode toolkit we might want to just replace it... but for now let's
workaround the issue by ignoring small distance animation.

Screenshot when hitting the breakpoint - buttons have wrong style and height:
 {F1272099178}

Reviewed By: zzl0

Differential Revision: D52667618

fbshipit-source-id: c183d7c60ce0e5da19afd39bf696cff89f5650a3