CommitTreeList: initial DAG-based commit list rendering
Summary:
Initial `<RenderDag />`-based graph rendering, off by default.

An internal config is added to use the new rendering logic:

  # render both
  sl config --edit isl.experimental-graph-renderer=2
  # render new
  sl config --edit isl.experimental-graph-renderer=1

Missing pieces:
- "You are moving here" spinner.
- Stack buttons.
- "Load more commits" button.

The selection background does not include the "o" circle because in complex cases
the "o" is not at the right most column, like B, C, D in this graph:

            o  E
      ╭─┬─┬─┤
      │ o │ │  D
      ╭─┴─╮ │
      │   o │  C
      │   ├─╯
      o   │  B
      ├───╯
      o  A

It might make sense to highlight the full row. But that'd be extra changes.

Reviewed By: muirdm, evangrayk

Differential Revision: D52646426

fbshipit-source-id: c568551c242ba94bf25bbeb4f72a5cec9c162c12