selection: treeWithPreviews => dagWithPreviews
Summary:
Now the dag.sortAsc order matches the linearizedCommitHistory, we can replace
`treeWithPreviews` with `dagWithPreviews`.

This is an intermediate state as it's not obvious why the order matches.
A future diff will replace the trees with dag rendering so there will be a
stronger guarantee that the order here matches the rendering.

Reviewed By: evangrayk

Differential Revision: D52522341

fbshipit-source-id: c671fc4b14bbde399fd212e6d130c295ceec36a5