adjust edenapi concurrency for fetching files.
Summary:
EdenFs uses the batch size 32. So, it is the most loggical to allow
concurrency up to the same size. Otherwise we just create an artificial 3x slow
down for these requests.

We have other protection like cachelib shards limit, etc, to keep out throuput
bounded, so it doesn't make much sense to configure this number to be 3x
smaller than the EdenFs batch.

https://fburl.com/scuba/mononoke_test_perf/bz3zda7t

We would improve a good amount of fetches.

https://fburl.com/scuba/mononoke_test_perf/8z6kz6le

In case, we would like to set it lower due to some reason, then we would need to lower the eden fs
batch size as well, so that they would go with a smaller batch at a time but to more
hosts concurrently to avoid artificial slow downs when we still have headroom
in throughput. The idea is that optimal performance is reached when these
two values are matched.

Differential Revision: D52833986

fbshipit-source-id: f4dc0b81c7a82b117b851ca38e9c6e84e0aee420