avoid unnecessary work on branch creation
Summary:
This fixes the problem highlighted in the previous commit in this stack.

Let's add some discovery logic that will try to traverse the repo in a search for first commit that's not yet imported. To make that fast we're doing exponential jumps. This works well for single-threaded branches. Doesn't work all that well for merge commits (as we follow only p1 history) but it won't be worse than what we're doing right now.

Differential Revision: D52841546

fbshipit-source-id: d56f685eef30405e9f0d4942778267030a8f79a1