doc: fix GitHub PR link in `getting-started.md`
The 'Getting Started' documentation page has a section about replacing the GitHub PR URL with reviewstack, but the link in Markdown is parsed as GitHub flavored Markdown and shows just `facebook/react#25506` instead of `https://github.com/facebook/react/pull/25506` which is slightly confusing. This change fixes that to the full intended URL is displayed.