Default to showing blame
Blame feature has been out for a while and seems functional generally. I often get questions about how to enable this.

Seems like it should be good to just enable by default, but if that conflicts with an internal blame or something I could try and ensure we set this setting in all our repos instead