cliparser: remove conversion of "foo --bar" to "foo-bar"
Summary:
Previously, "sl foo --bar" would actually run "sl foo-bar" if there was a command or alias named "foo-bar". This was added as a way to migrate commands such as "sl pull --rebase" that do "too much" to separate commands using an alias like `pull-rebase = !$HG pull && $HG rebase -d remote/master "$@"`.

This was confusing users since "sl pull --rebase -d master" was expanding the "pull-rebase" alias into the now invalid "sl pull && sl rebase -d remote/master -d master" (we tweaked the arg validation to be stricter recently). The confusion is exemplified by the workaround which was to switch the order of args to "sl pull -d master --rebase", defeating the alias.

And further note there is another level of confusion with HGPLAIN=1 since the alias won't take effect.

Anyway, remove the "foo --bar" -> "foo-bar" conversion since it is confusing and not really relied upon. This should be relatively safe because we already can be certain automation using HGPLAIN=1 is not relying on this functionality (because aliases are disabled).

Reviewed By: quark-zju

Differential Revision: D52877939

fbshipit-source-id: 16ade88d9a32efdf1789bd2cff4f2ef1b06137bb