cleanup symlink config in the daemon
Summary:
There is a typo in this config name. The config we use in the CLI that
determines the state of the rollout is  `experimental:windows-symlinks` instead
of `experimental:windows-symlinks-enabled`. This causes warning for the users.

Confirmed we don't use the `experimental:windows-symlinks-enabled` value
anywhere. and checked that we don't use `EdenConfig.windowsSymlinksEnabled`
value anywhere in the daemon.

I'm changing the default value to false too because the cli was asking the
daemon for the config and defaulting to false when the config was not set.
This ensures that symlinks stay disable.

Reviewed By: sggutier

Differential Revision: D52891836

fbshipit-source-id: cd25f16a73a6d2e2202dc16a4bfa8e052c8240b7